Progresss bar: Re-use clip duration from getting media info
As we already got the duration in seconds from running ffprobe to
get the media info, re-use it for the progress bar.

Besides simplifying the code, it also fixes some instances where
the progress bar would be stuck at 0% when parsing the ffmpeg
output for "Duration:" failed for some reason.