Confirmation to stop encoding
Show an alert asking for confirmation to abort the encoding
if the encoding has been running for more than 30 seconds.

Below that time, a confirmation alert would be more of an annoyance
than having to re-start the encoding if "Stop" was pressed
accidentally.

Show alert asynchronously to keep updating the progress bar.

Change alert text and button label if the encoding has finished
before the user clicked a button... :)