Show alert when adding a job with same output filename
We already ignore trying to add a job with the exact same ffmpeg
command line.
This change warns the user if the command line is different, but
the output file is already used by another job. That would
overwrite the result of that older job.

Unrelated: Keep fDuration containing the plain time string ("0:01:23")
and only add the UTF clock symbol to show up in the ColumnListView.
Otherwise, we'd have to strip the symbol before saving the "jobs" file.