Jobs: Add a sequential job number as 1st column
As a way to bring the list back into the order the jobs were added.
Otherwise it can get awkward: You could sort after Duration, then
when adding a new job it will appear at the bottom, but doesn't
actually obey the sorting order.

The job number isn't saved with the jobs file, but added when loading.
The sort order is also reset when loading the jobs file. It'll be in
the last used sorting order anyway, because jobs are saved from top
to bottom of the list.

The job numbering is fixed, there is no re-numbering if jobs are
removed. That complication doesn't seem to be warrented.