Fix initial crop spinner invocation
I'm not sure why, but when using SetWithoutInvoke() with the crop spinners,
the M_*CROP message won't get sent until you use the +/- buttons. That is,
if you just enter values in the top/bottom/left/right text boxes, it won't
trigger the drawing of the crop box or the _BuildLine().

Using SetValue() works, so let's just use that then...

Curious why the other spinners (video bitrate etc.) don't show this problem.